init kotpref works

+++ app/src/main/java/com/example/exchangerate/data/datasource/database/model/CurrencyRateEntity.kt
@@ -9,5 +9,5 @@ data class CurrencyRateEntity(
     @PrimaryKey
     val id: String,
     @ColumnInfo(name = "rate_in_usd")
-    val rateInUsd: Float
+    val rateInUsd: Double
 )
