+++ app/src/main/java/com/example/exchangerate/model/Currency.kt
@@ -3,5 +3,5 @@ package com.example.exchangerate.model
 data class Currency(
     val id: String,
     val name: String,
-    val rateInUsd: Float
+    val rateInUsd: Double
 )
